Split generate payment rollback helpers
